+++ CHANGELOG.rst
@@ -6,11 +6,17 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog`_,
 and this project adheres to `Semantic Versioning`_.
 
-`Unreleased`_
+`1.3.1`_ - 2020-07-09
 --------------------------
 Added
 '''''
 - VTK output for seed lists and polyhedral meshes.
+- Option to compute expected area of ellipse from area distribution.
+- Option to compute expected volume of ellipsoid from volume distribution.
+
+Fixed
+'''''
+- Error in verification module for 2D uniform random orientations.
 
 `1.3.0`_ - 2020-06-25
 --------------------------
@@ -113,7 +119,8 @@ Added
 
 .. LINKS
 
-.. _`Unreleased`: https://github.com/kip-hart/MicroStructPy/compare/v1.3.0...HEAD
+.. _`Unreleased`: https://github.com/kip-hart/MicroStructPy/compare/v1.3.1...HEAD
+.. _`1.3.1`: https://github.com/kip-hart/MicroStructPy/compare/v1.3.0...v1.3.1
 .. _`1.3.0`: https://github.com/kip-hart/MicroStructPy/compare/v1.2.2...v1.3.0
 .. _`1.2.2`: https://github.com/kip-hart/MicroStructPy/compare/v1.2.1...v1.2.2
 .. _`1.2.1`: https://github.com/kip-hart/MicroStructPy/compare/v1.2.0...v1.2.1